requirements.txt
gradio~=5.7
google-genai
pandas
openpyxl
httpx
beautifulsoup4
python-dotenv
requests
